CRTを利用するように修正
